refactor(export): extract shared blob-download helper from CSV export
Also exports formatCategoryPath so the upcoming JSON export can reuse the
same parent/child path formatting instead of duplicating it.
