+++ apps/web-next/src/utility/csvExport.ts
@@ -1,4 +1,5 @@
 import type { TransactionExportRow } from "./exportTransactions";
+import { downloadTextFile } from "./fileDownload";
 
 const CSV_HEADER = [
   "Date",
@@ -18,7 +19,7 @@ export const escapeCsvField = (
   return /[",\n\r]/.test(str) ? `"${str.replace(/"/g, '""')}"` : str;
 };
 
-const formatCategoryPath = (
+export const formatCategoryPath = (
   name: string | null,
   parentName: string | null
 ): string => {
@@ -43,14 +44,5 @@ export const buildCsv = (rows: string[][]): string => {
   return lines.join("\r\n") + "\r\n";
 };
 
-export const downloadCsv = (filename: string, content: string): void => {
-  const blob = new Blob([content], { type: "text/csv;charset=utf-8;" });
-  const url = URL.createObjectURL(blob);
-  const link = document.createElement("a");
-  link.href = url;
-  link.download = filename;
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-  URL.revokeObjectURL(url);
-};
+export const downloadCsv = (filename: string, content: string): void =>
+  downloadTextFile(filename, content, "text/csv;charset=utf-8;");
