+++ apps/web-next/src/utility/fileDownload.ts
@@ -1,0 +1,15 @@
+export const downloadTextFile = (
+  filename: string,
+  content: string,
+  mimeType: string
+): void => {
+  const blob = new Blob([content], { type: mimeType });
+  const url = URL.createObjectURL(blob);
+  const link = document.createElement("a");
+  link.href = url;
+  link.download = filename;
+  document.body.appendChild(link);
+  link.click();
+  document.body.removeChild(link);
+  URL.revokeObjectURL(url);
+};
